Renaming files to conform to PEP8
