+++ permuta/__init__.py
@@ -1,6 +1,6 @@
-from .Perm import *
-from .PermSet import *
-from .MeshPatt import MeshPatt, gen_meshpatts
+from .perm import *
+from .permset import *
+from .meshpatt import MeshPatt, gen_meshpatts
 from . import _perm_set
 from . import descriptors
 from . import interfaces
