+++ permuta/_perm_set/__init__.py
@@ -1,3 +1,3 @@
-from .PermSetBase import PermSetBase
+from .permset_base import PermSetBase
 from . import finite
 from . import unbounded
