+++ permuta/_perm_set/finite/__init__.py
@@ -1,3 +1,3 @@
-from .PermSetFinite import PermSetFinite
-from .PermSetFiniteSpecificLength import PermSetFiniteSpecificLength
-from .PermSetStatic import PermSetStatic
+from .permset_finite import PermSetFinite
+from .permset_finite_specificlength import PermSetFiniteSpecificLength
+from .permset_static import PermSetStatic
