+++ permuta/_perm_set/unbounded/__init__.py
@@ -1,3 +1,3 @@
-from .PermSetUnbounded import PermSetUnbounded
+from .permset_unbounded import PermSetUnbounded
 from . import all
 from . import described
