+++ permuta/_perm_set/unbounded/all/__init__.py
@@ -1,1 +1,1 @@
-from .PermSetAll import PermSetAll
+from .permset_all import PermSetAll
