+++ permuta/_perm_set/unbounded/all/permset_all.py
