+++ permuta/_perm_set/unbounded/described/avoiding/__init__.py
@@ -1,4 +1,4 @@
-from .Avoiding import Avoiding, AvoidingGeneric, AvoidingGenericSpecificLength
+from .avoiding import Avoiding, AvoidingGeneric, AvoidingGenericSpecificLength
 
 from . import avoiding_length_0
 #from . import avoiding_length_1
