+++ permuta/_perm_set/unbounded/described/avoiding/avoiding_length_0/__init__.py
@@ -1,1 +1,1 @@
-from .AvoidingEmpty import AvoidingEmpty
+from .avoiding_empty import AvoidingEmpty
