+++ permuta/_perm_set/unbounded/described/avoiding/avoiding_length_0/avoiding_empty.py
@@ -2,7 +2,7 @@ from permuta import Perm
 from permuta.descriptors import Basis
 from permuta._perm_set.finite import PermSetStatic
 
-from ..Avoiding import Avoiding
+from ..avoiding import Avoiding
 
 
 class AvoidingEmpty(Avoiding):
