+++ permuta/_perm_set/unbounded/permset_unbounded.py
@@ -1,4 +1,4 @@
-from ..PermSetBase import PermSetBase
+from ..permset_base import PermSetBase
 
 
 class PermSetUnbounded(PermSetBase):
