+++ permuta/descriptors/__init__.py
@@ -1,3 +1,3 @@
-from .Descriptor import Descriptor
-from .Basis import Basis
-from .Predicate import Predicate
+from .descriptor import Descriptor
+from .basis import Basis
+from .predicate import Predicate
