+++ permuta/descriptors/descriptor.py
