+++ permuta/interfaces/__init__.py
@@ -1,4 +1,4 @@
-from .Patt import Patt
-from .Flippable import Flippable
-from .Rotatable import Rotatable
-from .Shiftable import Shiftable
+from .patt import Patt
+from .flippable import Flippable
+from .rotatable import Rotatable
+from .shiftable import Shiftable
