+++ permuta/interfaces/flippable.py
