+++ permuta/interfaces/rotatable.py
