+++ permuta/interfaces/shiftable.py
